+++ psycopg/psycopg/server_cursor.py
@@ -216,7 +216,7 @@ class ServerCursor(ServerCursorMixin["Connection[Any]", Row], Cursor[Row]):
 
     @overload
     def __init__(
-        self: "ServerCursor[Row]",
+        self,
         connection: "Connection[Row]",
         name: str,
         *,
@@ -227,7 +227,7 @@ class ServerCursor(ServerCursorMixin["Connection[Any]", Row], Cursor[Row]):
 
     @overload
     def __init__(
-        self: "ServerCursor[Row]",
+        self,
         connection: "Connection[Any]",
         name: str,
         *,
@@ -357,7 +357,7 @@ class AsyncServerCursor(
 
     @overload
     def __init__(
-        self: "AsyncServerCursor[Row]",
+        self,
         connection: "AsyncConnection[Row]",
         name: str,
         *,
@@ -368,7 +368,7 @@ class AsyncServerCursor(
 
     @overload
     def __init__(
-        self: "AsyncServerCursor[Row]",
+        self,
         connection: "AsyncConnection[Any]",
         name: str,
         *,