+++ tests/test_cursor_server_async.py
@@ -6,6 +6,7 @@ from psycopg.pq import Format
 
 from .acompat import alist
 
+
 pytestmark = pytest.mark.crdb_skip("server-side cursor")
 
 
@@ -261,11 +262,12 @@ async def test_close_no_clobber(aconn):
             await cur.fetchall()
 
 
-async def test_warn_close(aconn, recwarn):
+async def test_warn_close(aconn, recwarn, gc_collect):
     recwarn.clear()
     cur = aconn.cursor("foo")
     await cur.execute("select generate_series(1, 10) as bar")
     del cur
+    gc_collect()
     assert ".close()" in str(recwarn.pop(ResourceWarning).message)